+++ internal/pkg/docker/dockerfile/dockerfile_test.go
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package exec
+package dockerfile
 
 import (
 	"fmt"
