+++ internal/pkg/exec/errors.go
@@ -3,28 +3,6 @@
 
 package exec
 
-import (
-	"fmt"
-)
-
-// ErrInvalidPort means that while there was a port provided, it was out of bounds or unparseable
-type ErrInvalidPort struct {
-	Match string
-}
-
-func (e ErrInvalidPort) Error() string {
-	return fmt.Sprintf("port represented at %s is invalid or unparseable", e.Match)
-}
-
-// ErrNoExpose means there were no documented EXPOSE statements in the given dockerfile.
-type ErrNoExpose struct {
-	Dockerfile string
-}
-
-func (e ErrNoExpose) Error() string {
-	return fmt.Sprintf("no EXPOSE statements in Dockerfile %s", e.Dockerfile)
-}
-
 // ErrSSMPluginNotExist means the ssm plugin is not installed.
 type ErrSSMPluginNotExist struct{}
